Updates some entrypoints

+++ brayns/network/messages/RemoveModelMessage.h
@@ -26,6 +26,6 @@
 namespace brayns
 {
 BRAYNS_JSON_OBJECT_BEGIN(RemoveModelMessage)
-BRAYNS_JSON_OBJECT_ENTRY(std::vector<size_t>, ids, "List of model ID to remove")
+BRAYNS_JSON_OBJECT_ENTRY(std::vector<uint32_t>, ids, "List of model ID to remove")
 BRAYNS_JSON_OBJECT_END()
 } // namespace brayns
